requirements.txt
appdirs==1.4.4
attrs==20.2.0
aws-requests-auth==0.4.1
boto3==1.12.3
botocore==1.15.31
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
clickclick==1.2.2
connexion==2.5.1
cycler==0.10.0
distlib==0.3.1
docutils==0.15.2
elasticsearch==7.0.0
elasticsearch-dsl==7.2.1
Faker==4.0.2
filelock==3.0.12
Flask==1.1.2
gpt-2-simple==0.7.1
grpcio==1.27.2
h5py==2.10.0
idna==2.6
inflection==0.5.1
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
json2html==1.3.0
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.2
numpy==1.16.2
openapi-spec-validator==0.2.9
pandas==1.1.2
pick==1.0.0
Pillow==7.2.0
pony==0.7.13
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.9.27
requests==2.22.0
s3transfer==0.3.3
six==1.15.0
text-unidecode==1.3
toposort==1.5
tqdm==4.50.0
urllib3==1.22
virtualenv==20.0.31
Werkzeug==1.0.1
